fix: Add specific directories to feed into fzf
The start-session command was not reliably starting a tmux session on first use when the desired directory's full name was not typed out in fzf. However, the command would work after 3 or 4 attempts at starting a session. Typing the directory's full name would always start a tmux session on the first try. Limiting the directories that are fed into fzf seems to solve the issue.